Add test_match_ayon_hierarchy tests.
